# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: RoomFlow API
  version: 1.0.0
  contact:
    name: RoomFlow Team
  description: |-
    RoomFlow 会议室预约管理系统 REST API 契约（OpenAPI 3.0.3）。

    本文件是前后端唯一共享契约：后端 Controller 与前端 api/types 均以此为准。
    配套说明文档（错误码表、分页约定、权限矩阵、时区约定）见 `docs/api/README.md`。

    通用约定：
    - 统一响应体 `Result<T>`：`{code:int, message:string, data:T}`。`code=0` 成功；非 0 为业务错误码，同时携带匹配的 HTTP 状态码。
    - 时间：全部时间字段（请求与响应）均为北京时间，ISO 8601 格式且必须显式携带 `+08:00` 偏移，例：`2026-09-16T14:00:00+08:00`。`format: date` 的查询参数为北京时间日历日 `YYYY-MM-DD`。
    - 会议时段为 `[startTime, endTime)` 半开区间；同一房间的相邻时段允许首尾相接（前一会的 endTime 可等于后一会的 startTime）。
    - 时间规则：`startTime` 不得早于当前时间、不得晚于当前时间 +72h（滚动窗口只约束开始时间）；`startTime` 与 `endTime` 均须按 15 分钟对齐（秒=0 且分钟 %15=0）；时长为 15min–24h（含两端）；允许跨天；`endTime` 可超出 72h 预约窗口。
    - 权限：每个端点以 `x-permission` 标注，取值 `PUBLIC`（无需登录）/ `REFRESH_TOKEN`（仅凭请求体中的 Refresh Token）/ `AUTHENTICATED`（任意已登录用户）/ `ADMIN` / `ORGANIZER_OR_ADMIN`（会议发起人或管理员）/ `SELF`（仅操作本人资源）。全局默认要求 Bearer JWT；公开端点显式声明 `security: []`。
    - 每个端点可能返回的业务错误码以 `x-error-codes` 枚举，与 README 错误码表一一对应。
    - 分页：`page`（从 1 起）/ `size`（默认 10，最大 100），分页响应为 `PageResult<T> = {records, total, page, size}`。
servers:
  - url: http://localhost:8080
    description: 本地开发环境
...
security:
  - bearerAuth: []
paths:
...
  /api/rooms:
    get:
      tags:
        - rooms
      operationId: listRooms
      summary: 会议室列表
      description: |-
        返回全部会议室（含已停用），按 id 升序，不分页（房间是少量有限资源）。
        可选 `enabled` 筛选启用状态；缺省返回全部。
      x-permission: AUTHENTICATED
      x-error-codes: [40001, 40101, 40102, 50000]
      parameters:
        - name: enabled
          in: query
          required: false
          schema:
            type: boolean
          description: 按启用状态筛选；true=仅启用，false=仅停用，缺省=全部
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultRoomVOList'
              example:
                code: 0
                message: success
                data:
                  - id: 1
                    name: 301会议室
                    location: 3楼东侧
                    capacity: 8
                    equipment: [PROJECTOR, WHITEBOARD]
                    enabled: true
                    createdAt: '2026-09-15T09:00:00+08:00'
                    updatedAt: '2026-09-15T09:00:00+08:00'
                  - id: 2
                    name: 302会议室
                    location: 3楼西侧
                    capacity: 12
                    equipment: [PROJECTOR, VIDEO_CONFERENCE]
                    enabled: true
                    createdAt: '2026-09-15T09:00:00+08:00'
                    updatedAt: '2026-09-15T09:00:00+08:00'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags:
        - rooms
      operationId: createRoom
      summary: 创建会议室
      description: 创建会议室，enabled 默认为 true。仅 ADMIN。
      x-permission: ADMIN
      x-error-codes: [40001, 40002, 40101, 40102, 40301, 50000]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoomRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultRoomVO'
              example:
                code: 0
                message: success
                data:
                  id: 4
                  name: 多功能厅B
                  location: 5楼
                  capacity: 20
                  equipment: [PROJECTOR, WHITEBOARD, VIDEO_CONFERENCE, PHONE]
                  enabled: true
                  createdAt: '2026-09-15T14:00:00+08:00'
                  updatedAt: '2026-09-15T14:00:00+08:00'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/rooms/{id}:
    get:
      tags:
        - rooms
      operationId: getRoom
      summary: 会议室详情
      description: 按 ID 查询会议室（含已停用房间）。
      x-permission: AUTHENTICATED
      x-error-codes: [40001, 40101, 40102, 40401, 50000]
      parameters:
        - $ref: '#/components/parameters/PathId'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultRoomVO'
              example:
                code: 0
                message: success
                data:
                  id: 1
                  name: 301会议室
                  location: 3楼东侧
                  capacity: 8
                  equipment: [PROJECTOR, WHITEBOARD]
                  enabled: true
                  createdAt: '2026-09-15T09:00:00+08:00'
                  updatedAt: '2026-09-15T09:00:00+08:00'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    put:
      tags:
        - rooms
      operationId: updateRoom
      summary: 修改会议室
      description: |-
        全量更新会议室信息（PUT 语义）。仅 ADMIN。
        location、equipment 省略时分别视为 null、空数组。
        注意：本端点不修改 enabled，启停请使用 PATCH /api/rooms/{id}/status。
      x-permission: ADMIN
      x-error-codes: [40001, 40002, 40101, 40102, 40301, 40401, 50000]
      parameters:
        - $ref: '#/components/parameters/PathId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoomRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultRoomVO'
              example:
                code: 0
                message: success
                data:
                  id: 1
                  name: 301会议室（翻新版）
                  location: 3楼东侧
                  capacity: 10
                  equipment: [PROJECTOR, WHITEBOARD, PHONE]
                  enabled: true
                  createdAt: '2026-09-15T09:00:00+08:00'
                  updatedAt: '2026-09-15T15:00:00+08:00'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags:
        - rooms
      operationId: deleteRoom
      summary: 停用会议室（软删除）
      description: |-
        软删除会议室：将 enabled 置为 false，不物理删除行，保留历史会议的 room_id 关联完整性。
        停用时校验：若该房间存在 status=ACTIVE 的未结束会议，拒绝并返回 409 + code=40907，需先处理相关会议。
        幂等：对已停用房间重复调用返回 200。仅 ADMIN。
      x-permission: ADMIN
      x-error-codes: [40001, 40101, 40102, 40301, 40401, 40907, 50000]
      parameters:
        - $ref: '#/components/parameters/PathId'
      responses:
        '200':
          description: 停用成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultVoid'
              example:
                code: 0
                message: success
                data: null
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/rooms/{id}/status:
    patch:
      tags:
        - rooms
      operationId: updateRoomStatus
      summary: 启用 / 停用会议室
      description: |-
        设置 enabled 标志。仅 ADMIN。
        enabled=false（停用）时校验：存在 status=ACTIVE 的未结束会议则拒绝（409 + code=40907）。
        enabled=true（启用）无附加校验。
      x-permission: ADMIN
      x-error-codes: [40001, 40002, 40101, 40102, 40301, 40401, 40907, 50000]
      parameters:
        - $ref: '#/components/parameters/PathId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoomStatusRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultRoomVO'
              example:
                code: 0
                message: success
                data:
                  id: 1
                  name: 301会议室
                  location: 3楼东侧
                  capacity: 8
                  equipment: [PROJECTOR, WHITEBOARD]
                  enabled: false
                  createdAt: '2026-09-15T09:00:00+08:00'
                  updatedAt: '2026-09-15T16:00:00+08:00'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/rooms/{id}/availability:
    get:
      tags:
        - rooms
      operationId: getRoomAvailability
      summary: 查询房间占用时段
      description: |-
        返回指定房间在指定北京时间日历日范围内、与范围有交集的全部 status=ACTIVE 会议占用时段
        （`occupiedSlots`，按 startTime 升序），由前端用“日期范围 − 占用时段”计算空闲时段。
        设计理由：只暴露已占用时段这一事实数据，空闲区间的生成规则（粒度、边界）由前端决定，
        避免前后端对“空闲段”边界定义不一致。
        已停用房间（enabled=false）仍可查询其占用时段（历史会议保留关联）。
        缺少必填查询参数 startDate 时返回 400 + code=40002。
      x-permission: AUTHENTICATED
      x-error-codes: [40001, 40002, 40101, 40102, 40401, 50000]
      parameters:
        - $ref: '#/components/parameters/PathId'
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date
          description: 范围起始日，北京时间日历日 YYYY-MM-DD
          example: '2026-09-16'
        - name: endDate
          in: query
          required: false
          schema:
            type: string
            format: date
          description: 范围结束日（含），北京时间日历日；缺省=startDate；不得早于 startDate 且跨度 ≤ 7 天，否则 40001
          example: '2026-09-16'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultRoomAvailabilityVO'
              example:
                code: 0
                message: success
                data:
                  roomId: 1
                  startDate: '2026-09-16'
                  endDate: '2026-09-16'
                  occupiedSlots:
                    - meetingId: 101
                      startTime: '2026-09-16T10:00:00+08:00'
                      endTime: '2026-09-16T11:30:00+08:00'
                    - meetingId: 102
                      startTime: '2026-09-16T14:00:00+08:00'
                      endTime: '2026-09-16T15:00:00+08:00'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |-
        将 Access Token 放入请求头：`Authorization: Bearer <accessToken>`。
        未携带 / 签名无效 / 已被撤销（logout 或账号禁用后）→ 401 + code=40101；
        已过期 → 401 + code=40102（前端应调用 POST /api/auth/refresh 换新）。
  parameters:
    PathId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
      description: 资源 ID
...
  responses:
    BadRequest:
      description: 参数校验失败（code=40001，data.fieldErrors 给出字段级错误）或请求格式错误（code=40002，如 JSON 解析失败 / 类型不匹配）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResult'
          example:
            code: 40001
            message: 参数校验失败
            data:
              fieldErrors:
                - field: startTime
                  message: 必须按 15 分钟对齐
    Unauthorized:
      description: 未认证或凭证无效（code=40101 未携带/无效/已撤销，40102 Token 过期，40103 用户名或密码错误，40104 Refresh Token 无效/过期/已撤销）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResult'
          example:
            code: 40101
            message: 未认证或 Token 无效
            data: null
    Forbidden:
      description: 已认证但无权限（code=40301 无权限，40302 账号已禁用）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResult'
          example:
            code: 40301
            message: 无权限执行该操作
            data: null
    NotFound:
      description: 资源不存在或对当前用户不可见（code=40401）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResult'
          example:
            code: 40401
            message: 资源不存在
            data: null
    Conflict:
      description: 业务规则冲突，code 取值见该端点 x-error-codes（40901–40909）与 README 错误码表
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResult'
          example:
            code: 40902
            message: 会议时间与该房间已有会议冲突
            data: null
    InternalError:
      description: 服务器内部错误（code=50000），响应不包含堆栈或内部细节
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResult'
          example:
            code: 50000
            message: 服务器内部错误
            data: null
  schemas:
    FieldError:
      type: object
      required:
        - field
        - message
      properties:
        field:
          type: string
          description: 校验失败的字段名（JSON 字段路径）
        message:
          type: string
          description: 字段级错误信息（中文）
    ErrorResult:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
          description: 非 0 业务错误码，见 README 错误码表
        message:
          type: string
          description: 人类可读错误信息（中文），不含敏感内部细节
        data:
          type: object
          nullable: true
          description: 通常为 null；code=40001 时携带 fieldErrors 字段级错误列表
          properties:
            fieldErrors:
              type: array
              items:
                $ref: '#/components/schemas/FieldError'
    Result:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
          description: 0=成功
        message:
          type: string
          description: 提示信息（中文），成功时为 success 或操作描述
        data:
          description: 业务数据，类型见各 ResultXxx 包装 schema；无数据操作时为 null
    ResultVoid:
      allOf:
        - $ref: '#/components/schemas/Result'
        - type: object
          properties:
            data:
              type: object
              nullable: true
              description: 固定为 null
...
    ResultRoomVO:
      allOf:
        - $ref: '#/components/schemas/Result'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/RoomVO'
    ResultRoomVOList:
      allOf:
        - $ref: '#/components/schemas/Result'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/RoomVO'
    ResultRoomAvailabilityVO:
      allOf:
        - $ref: '#/components/schemas/Result'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/RoomAvailabilityVO'
...
    Equipment:
      type: string
      enum:
        - PROJECTOR
        - WHITEBOARD
        - VIDEO_CONFERENCE
        - PHONE
      description: 固定设备选项：PROJECTOR=投影仪，WHITEBOARD=白板，VIDEO_CONFERENCE=视频会议，PHONE=电话
    RoomVO:
      type: object
      required:
        - id
        - name
        - location
        - capacity
        - equipment
        - enabled
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          maxLength: 100
        location:
          type: string
          maxLength: 200
          nullable: true
        capacity:
          type: integer
          format: int32
          minimum: 1
          maximum: 100
        equipment:
          type: array
          uniqueItems: true
          items:
            $ref: '#/components/schemas/Equipment'
          description: 设备列表，可为空数组
        enabled:
          type: boolean
          description: true=启用可预约；false=停用（软删除，历史关联保留）
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateRoomRequest:
      type: object
      required:
        - name
        - capacity
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        location:
          type: string
          maxLength: 200
          nullable: true
        capacity:
          type: integer
          format: int32
          minimum: 1
          maximum: 100
        equipment:
          type: array
          uniqueItems: true
          items:
            $ref: '#/components/schemas/Equipment'
          description: 省略或空数组表示无设备
    UpdateRoomRequest:
      type: object
      description: PUT 全量更新；location 省略视为 null，equipment 省略视为空数组；不含 enabled（走 status 端点）
      required:
        - name
        - capacity
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        location:
          type: string
          maxLength: 200
          nullable: true
        capacity:
          type: integer
          format: int32
          minimum: 1
          maximum: 100
        equipment:
          type: array
          uniqueItems: true
          items:
            $ref: '#/components/schemas/Equipment'
    RoomStatusRequest:
      type: object
      required:
        - enabled
      properties:
        enabled:
          type: boolean
          description: true=启用；false=停用（存在 ACTIVE 会议时拒绝，40907）
    TimeSlot:
      type: object
      description: 已占用时段，`[startTime, endTime)` 半开区间
      required:
        - meetingId
        - startTime
        - endTime
      properties:
        meetingId:
          type: integer
          format: int64
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
    RoomAvailabilityVO:
      type: object
      description: |-
        房间占用查询结果。occupiedSlots 为与查询日期范围有交集的 status=ACTIVE 会议时段，
        按 startTime 升序。空闲时段 = 查询范围（北京时间日历日 00:00–24:00）减去占用时段，由前端计算。
      required:
        - roomId
        - startDate
        - endDate
        - occupiedSlots
      properties:
        roomId:
          type: integer
          format: int64
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        occupiedSlots:
          type: array
          items:
            $ref: '#/components/schemas/TimeSlot'
...